charts: fix helm lint errors
Removed the ability to specify .Values.secrets.site_key.create to
control the creation of the site_key secret. Instead, it is determined
by the presence of .Values.secrets.site_key.value :
if present, create the secret. if not, use the pre-existing secret
supplied in .Values.secrets.site_key.name.
If both are empty, work in offline mode.