hooks: added pre delete hook
Upon "helm uninstall" to the s1 chart, this hook will start a job.
the job will run a pod using the s1-helper image.
The container inside the pod will use kubectl to execute on each agent
pod the "sentinelctl control uninstall" command which is now mandatory
to uninstall the agent.

also, added the permission to execute (kubectl exec) to the service
account.