all: fix injected agent uninstall in severless mode
Helper, running on Fargate pod, creates secrets and configmaps
in each namespace which has injected agents. On uninstall, helm
for some reason deletes the cluster role and bindings before the helper
shuts down, removes the secrets and updates the configmap. Maybe it
takes a while for the helper to shutdown on Fargate and in the meanwhile
helm removes the rbac so the helper fails to update the configmap with
uninstall status. Therfore, we added annotation to the rbac for keeping
them.