charts: s1-agent: support argocd post delete hook
For argocd deployments, support an argocd post delete hook.
We define a job that will run after a delete app occurred in argocd,
use the preserved daemonset, statefulset and serviceaccount
in order to trigger uninstall on all our agents and helper.
Then, we remove these 3 resources.

This approach is similar to helm's pre-delete-hook we have.