charts: agent, helper: add secrets cksum annotation
We are keeping the sha256 hash checksum of the secrets.yaml file
as an annotation in both the statefulset and daemonset, in order
to enforce a re-rollout of both of them on a helm upgrade,
so they both use the same helper token and can communicate with
each other.

Before - on a `helm upgrade` run, depending on what's the change,
it could be that only the agent is re-deployed, and since we
generate a new helper token secret on each upgrade, it would use
the newly generated one, while the helper is not re-deployed
and uses the old helper token secret value.

Also, remove timestamp annotation that we used as a workaround
for such situations