charts: s1-agent: templates: fix uninstall hooks for openshift
Consists of two main fixes:
1. Use the same securityContext for the job pod as we use in the
helper pod in statefulset.yaml. Otherwise - it will run with a
random user, and not user 1000.

2. ArgoCD only - make sure the SCC we create for openshift is
not deleted on uninstall, and remains while the post-delete-hook
runs. This will allow us to execute commands on s1-helper.
We also make sure we delete it when we finish.